refactor(store): align task_proposal snapshot migration with neighborhood
Neighborhood-alignment cleanup of Task 1's migration — no behavior change,
all 3 migration tests still pass. Three fixes against the sibling
migrations in the same file:

- Return `anyhow::Result<()>` (the file's `Result` alias) instead of
  `rusqlite::Result<()>`. Matches every other migration fn in the file;
  `?` on rusqlite errors auto-converts through anyhow.
- Revert `pub fn` back to crate-private `fn`. No external caller needs
  it now that the tests live inline.
- Wrap the table rebuild in `PRAGMA foreign_keys=OFF; BEGIN; ...;
  COMMIT; PRAGMA foreign_keys=ON;` to mirror
  `migrate_drop_agents_status_and_session_id_columns`. Currently safe
  without it, but the bracketing is the neighbor's template and future
  readers shouldn't have to re-prove that on every rebuild.

Also moves the 3 snapshot-migration tests + their v1-schema fixture
from `tests/store_tests.rs` into a new inline
`#[cfg(test)] mod snapshot_migration_tests` next to the existing
`backfill_tests` module — matches how migration tests are organized in
this file and lets the tests call the now-private migration fn directly.

Net test count unchanged (411): lib tests +3 (344 -> 347), store_tests
-3 (67 -> 64).
