+++ tests/store_tests.rs
@@ -2471,218 +2471,3 @@ fn dismiss_task_proposal_posts_updated_task_proposal_message() {
     assert_eq!(v["status"], "dismissed");
     assert!(v["taskNumber"].is_null());
 }
-
-// --- Task proposal snapshot column migration (v2) ---
-//
-// These tests exercise the migration that adds the 5 snapshot columns + the
-// `source_message_id` pointer to `task_proposals`. They bootstrap a bare
-// v1-only schema (no v2 columns) and then run the migration explicitly so we
-// can observe pre- and post-migration state. `make_store()` is unsuitable
-// because `Store::open` applies every migration up front.
-
-/// Bootstrap a Connection with the pre-v2 `task_proposals` schema (plus the
-/// minimum referenced tables). Mirrors what a DB opened before this task
-/// landed would look like.
-fn open_test_db_v1_schema() -> Connection {
-    let conn = Connection::open_in_memory().unwrap();
-    conn.execute_batch(
-        "CREATE TABLE channels (
-            id TEXT PRIMARY KEY,
-            name TEXT UNIQUE NOT NULL,
-            description TEXT,
-            channel_type TEXT NOT NULL DEFAULT 'channel',
-            archived INTEGER NOT NULL DEFAULT 0,
-            created_at TEXT NOT NULL DEFAULT (datetime('now')),
-            parent_channel_id TEXT REFERENCES channels(id)
-         );
-         CREATE TABLE messages (
-            id TEXT PRIMARY KEY,
-            channel_id TEXT NOT NULL,
-            sender_name TEXT NOT NULL,
-            sender_type TEXT NOT NULL,
-            sender_deleted INTEGER NOT NULL DEFAULT 0,
-            content TEXT NOT NULL,
-            created_at TEXT NOT NULL DEFAULT (datetime('now')),
-            seq INTEGER NOT NULL,
-            UNIQUE(channel_id, seq)
-         );
-         CREATE TABLE task_proposals (
-            id TEXT PRIMARY KEY,
-            channel_id TEXT NOT NULL REFERENCES channels(id) ON DELETE CASCADE,
-            proposed_by TEXT NOT NULL,
-            title TEXT NOT NULL,
-            status TEXT NOT NULL DEFAULT 'pending'
-                CHECK (status IN ('pending', 'accepted', 'dismissed')),
-            created_at TEXT NOT NULL DEFAULT (datetime('now')),
-            accepted_task_number INTEGER,
-            accepted_sub_channel_id TEXT REFERENCES channels(id) ON DELETE SET NULL,
-            resolved_by TEXT,
-            resolved_at TEXT
-         );
-         CREATE INDEX idx_task_proposals_channel_status ON task_proposals(channel_id, status);
-         INSERT INTO channels (id, name) VALUES ('ch', 'eng');",
-    )
-    .unwrap();
-    conn
-}
-
-#[test]
-fn migration_add_snapshot_columns_is_idempotent() {
-    let conn = open_test_db_v1_schema();
-    chorus::store::migrations::migrate_add_task_proposal_snapshot_columns(&conn).unwrap();
-    chorus::store::migrations::migrate_add_task_proposal_snapshot_columns(&conn).unwrap();
-    let n: i64 = conn
-        .query_row(
-            "SELECT COUNT(*) FROM pragma_table_info('task_proposals') WHERE name = 'snapshot_content'",
-            [],
-            |r| r.get(0),
-        )
-        .unwrap();
-    assert_eq!(n, 1);
-}
-
-#[test]
-fn migrated_db_rejects_partial_snapshot_insert() {
-    let conn = open_test_db_v1_schema();
-    // Pre-migration row with NULL snapshot (survives migration).
-    conn.execute(
-        "INSERT INTO task_proposals (id, channel_id, proposed_by, title, status, created_at)
-         VALUES ('old-prop', 'ch', 'claude', 't', 'pending', datetime('now'))",
-        [],
-    )
-    .unwrap();
-    chorus::store::migrations::migrate_add_task_proposal_snapshot_columns(&conn).unwrap();
-
-    // Attempt a partial snapshot (content set, other snapshot fields NULL).
-    // The CHECK must reject it on migrated DBs just as on fresh installs.
-    let res = conn.execute(
-        "INSERT INTO task_proposals (id, channel_id, proposed_by, title, status, created_at,
-            snapshot_content) VALUES ('p1', 'ch', 'claude', 't', 'pending',
-            datetime('now'), 'partial')",
-        [],
-    );
-    assert!(
-        res.is_err(),
-        "migrated DB must enforce the same CHECK as fresh DB"
-    );
-}
-
-#[test]
-fn legacy_v1_row_survives_migration_with_all_columns_preserved() {
-    let conn = open_test_db_v1_schema();
-    // Seed a v1 row exercising every column including terminal-state fields.
-    // `accepted_sub_channel_id` references channels(id), so seed that row too.
-    conn.execute(
-        "INSERT INTO channels (id, name) VALUES ('sub-ch-7', 'eng__task-7')",
-        [],
-    )
-    .unwrap();
-    conn.execute(
-        "INSERT INTO task_proposals (
-            id, channel_id, proposed_by, title, status, created_at,
-            accepted_task_number, accepted_sub_channel_id,
-            resolved_by, resolved_at
-         ) VALUES (
-            'legacy-1', 'ch', 'claude', 'legacy title', 'accepted',
-            '2026-04-20T10:00:00Z', 7, 'sub-ch-7',
-            'alice', '2026-04-20T10:05:00Z'
-         )",
-        [],
-    )
-    .unwrap();
-
-    chorus::store::migrations::migrate_add_task_proposal_snapshot_columns(&conn).unwrap();
-
-    #[allow(clippy::type_complexity)]
-    let (
-        channel_id,
-        proposed_by,
-        title,
-        status,
-        created_at,
-        task_no,
-        sub_ch,
-        resolved_by,
-        resolved_at,
-        src_msg,
-        snap_content,
-        snap_sender,
-        snap_sender_type,
-        snap_created,
-        snapped_at,
-    ): (
-        String,
-        String,
-        String,
-        String,
-        String,
-        Option<i64>,
-        Option<String>,
-        Option<String>,
-        Option<String>,
-        Option<String>,
-        Option<String>,
-        Option<String>,
-        Option<String>,
-        Option<String>,
-        Option<String>,
-    ) = conn
-        .query_row(
-            "SELECT channel_id, proposed_by, title, status, created_at,
-                    accepted_task_number, accepted_sub_channel_id,
-                    resolved_by, resolved_at,
-                    source_message_id, snapshot_content, snapshot_sender_name,
-                    snapshot_sender_type, snapshot_created_at, snapshotted_at
-             FROM task_proposals WHERE id = 'legacy-1'",
-            [],
-            |r| {
-                Ok((
-                    r.get(0)?,
-                    r.get(1)?,
-                    r.get(2)?,
-                    r.get(3)?,
-                    r.get(4)?,
-                    r.get(5)?,
-                    r.get(6)?,
-                    r.get(7)?,
-                    r.get(8)?,
-                    r.get(9)?,
-                    r.get(10)?,
-                    r.get(11)?,
-                    r.get(12)?,
-                    r.get(13)?,
-                    r.get(14)?,
-                ))
-            },
-        )
-        .unwrap();
-
-    // Every v1 field preserved:
-    assert_eq!(channel_id, "ch");
-    assert_eq!(proposed_by, "claude");
-    assert_eq!(title, "legacy title");
-    assert_eq!(status, "accepted");
-    assert_eq!(created_at, "2026-04-20T10:00:00Z");
-    assert_eq!(task_no, Some(7));
-    assert_eq!(sub_ch.as_deref(), Some("sub-ch-7"));
-    assert_eq!(resolved_by.as_deref(), Some("alice"));
-    assert_eq!(resolved_at.as_deref(), Some("2026-04-20T10:05:00Z"));
-    // v2 snapshot columns absent, consistent with legacy rows:
-    assert!(src_msg.is_none());
-    assert!(snap_content.is_none());
-    assert!(snap_sender.is_none());
-    assert!(snap_sender_type.is_none());
-    assert!(snap_created.is_none());
-    assert!(snapped_at.is_none());
-
-    // Index preserved:
-    let idx_count: i64 = conn
-        .query_row(
-            "SELECT COUNT(*) FROM sqlite_master
-             WHERE type = 'index' AND name = 'idx_task_proposals_channel_status'",
-            [],
-            |r| r.get(0),
-        )
-        .unwrap();
-    assert_eq!(idx_count, 1);
-}
